display course settings page in IE9; fixes #8620
the course settings page was just loading as a canvas shell around a blank
area. after debugging, the deepest I was able to go in the stack trace was to
line 3241 of jquery-1.7.2.js, which in this case would be a call to the browser
native select() function on a textarea which is not visible. the call stems
from user_lists.js, which is where we'll just route around the issue by only
calling select if the element is visible.

test plan:
- make sure the course settings page loads on IE9
- make sure you can still add students from the page
- make sure the new course wizard still works, and the page where you add users
  correctly pre-selects the example text that populates the textarea

Change-Id: I5def08007d2a2f8f6f91beb1d4f85e8b174a772a
Reviewed-on: https://gerrit.instructure.com/10988
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Ryan Florence <ryanf@instructure.com>