add collection items to the activity stream
When a new item is created, all users following the collection will get
an activity stream post, except for the user who created the item.

This stream item type doesn't show up in the existing dashboard, but is
returned by the API.

test plan: use the api to create two users, and a collection, and have
them both follow the collection. Have user A create an item in the
collection. Verify that user B gets a stream item, but user A doesn't.

Change-Id: I84b6e537ce913e17e9480074ecce08bc99f3e14b
Reviewed-on: https://gerrit.instructure.com/11205
Reviewed-by: Brian Palmer <brianp@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>