Fixes invalid context urls for group discussions in unlocked modules. Refs #9161
When a discussion is a group assignment in a module that is not yet unlocked
because other prerequisite course modules aren't completed yet, the page
that shows why the current discussion is locked was generating invalid URL
links which caused a page error. This fixes the URL generation to give a
more helpful explanation for why it is locked.

Testing Plan:
=============
* Create a course modules where the second requires the first to be completed.
* Create a group of at least one student
* Create a discussion assignment for the 2nd module and make it a group assignment
* When the user tries to access the discussion before completing 1st module, they
  should see a sensible page explaining they need to complete the first module.
  Previously this would be an error page.

Change-Id: I90ee144ba4a809122d28caa5732ade3aad44516d
Reviewed-on: https://gerrit.instructure.com/11845
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Joe Tanner <joe@instructure.com>