new profile page closes #9369
users have a new profile page that displays a
custom bio, links, and contact methods as well as
how the viewing user knows them.

this profile can be a part of any page and is
displayed on the user in a course pages as well.

profiles are found at /about/:user_id

Change-Id: I3339144135d67415af9068d18776b691320c1938
Reviewed-on: https://gerrit.instructure.com/12298
Reviewed-by: Jon Jensen <jon@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>