fix jumping flash notices, fixes #9629
fix issue where flash notice is pinned to the left while it slides down
and then jumps to the middle. there seems to be a difference in how
browsers treat auto margins wrt widths and nested relative elements. we
just make sure everything that might need it gets an auto margin

test plan:
1. do something that generates a flash notice or error (e.g. send a
   message in conversations)
2. it should slide down from the top middle
3. this should work in all supported browsers (ie/ff/chrome/etc)

Change-Id: I6835b932bf057ec6083899a94fefb680b6836419
Reviewed-on: https://gerrit.instructure.com/12585
Reviewed-by: Ryan Shaw <ryan@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>