only show "use for grading" options to teachers
fixes: #10080

a student can create a discussion topic but cannot
create an assignment. the backend worked and would
not let you, but the fronted was just showing the
checkbox when it should not have.

test plan:
create a discussion topic
edit an existing discussion topic that is used
for grading

as a teacher: make sure you can edit assignment
settings on both

as a student: make sure you dont see the option
to use for grading on either

Change-Id: I937e0f4a14964ccdd633988e2d3f5db753ed5b77
Reviewed-on: https://gerrit.instructure.com/13071
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Jon Jensen <jon@instructure.com>