fixup "skip to content" link in webkit.
in ab97213e we added js to handle focusing #content
but we also added tab index="-1" to #content. this
made it so if you click anywhere on the top half of
the screen on the login screen it would focus the whole
top half as if it was a link.

Change-Id: Iea925e02da80a07633fcf1c3f809e1c0aaa49023
Reviewed-on: https://gerrit.instructure.com/13634
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Ryan Shaw <ryan@instructure.com>