Fix where Upcoming Events that show due on "Tomorrow" when it should be "Today". Fixes #10931
There was a caching problem where 2 assignments can have the same
due date but one can display "Tomorrow" and the other "Today". Both
assignments were due at 11:59pm on the same day.

The issue was related to caching the display with no expiration. The
cache is linked to the assignment and the user's permissions. The
cache key was updated to use Date.today so it will get recomputed
after crossing midnight and the key now also uses the Time.zone.name
instead of the Time.zone.utc_offset as that can create incorrect matches
when two users cross date boundaries.

Testing Notes:
==========
Testing locally may not be possible without having caching enabled
and then changing the machine date/time. The timezone offset is
stored as part of the cache, so changing timezone won't enable
it to be reproduced either.

I don't think a reproducible situation can be created easily unfortunately.

Change-Id: I5cdb12795de23a3e121aadd32c17ccb3a3016bd3
Reviewed-on: https://gerrit.instructure.com/14243
Reviewed-by: Jacob Fugal <jacob@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>