fix bulk private message attachment regression, fixes #11277
ensure that the attachment associations get created for each bulk private
message. previously we would explicitly create each message, and thus tell
it which attachments it should have. now we create a single message (the
root one) and do the other ones in a delayed job. since we were just
cloning the root message, the associations were not being copied

fortunately, we have all the data, so we can restore any associations that
should have gotten created

test plan:
1. send a bulk private message to a bunch of people, specifying one or
   more attachments
2. ensure that the attachment appears in each created/updated conversation
   (for both sender and recipient)

Change-Id: I1a72b3ef0a047a413b84e780811b84b298c6ec57
Reviewed-on: https://gerrit.instructure.com/14474
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Zach Pendleton <zachp@instructure.com>