[#5773] ruby 1.9.3 spec fixes
[#5773] Fix selenium/jquery_ui_spec.rb

encoding issue for strings in the spec file, 
adding the "# encoding: UTF-8" comment solved it


[#5773] Fix selenium/i18n_js_spec.rb

This was failing on string equivalency for 
comparing JS date formatting to Ruby date 
formatting.  The reason was ruby 1.9 capitalizes 
the VMS date (%v) slightly differently 
(it runs all caps).  upcasing both strings 
allows a fair comparison.  Change made to the 
specfile, around line 70.

[#5773] Fix selenium/grades_spec.rb

Two of the failures here were due to using the 
"format_message" method from the TextHelper class 
to output directly to a view.  Arrays of strings 
in ruby 1.8 print out without their brackets, 
whereas in 1.9 the brackets are included.  
Using "join" immediately upon the output seems to 
fix it [EDIT: Per Brian P, used "first" instead
of "join"]. There was one other failure related to 
grade modification, where the editing was not 
occurring properly due to the "blur" event on the 
grade entry screen not being triggered when the 
javascript driver sent the "enter" keystroke.  I 
was able to get the thing to pass by using a 
direct javascript invocation from the 
selenium spec instead.

[#5773] Fix selenium/profile_spec.rb

encoding issue for a string in the spec file, 
adding the "# encoding: UTF-8" comment solved it

Change-Id: Idef84c88824dd39fbde99cf1efe095178a3a12e1
Reviewed-on: https://gerrit.instructure.com/14726
Reviewed-by: Brian Palmer <brianp@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>