tweak to go along with outcomes refactor
the recent outcomes refactor included temporary shims to let old code in
plugins work. this code used to be in a plugin, but is now that on
master, so we can fix it to move off the shims. once this and like
commits in other plugins have been submitted, the shims included with
the refactor can be removed.

test-plan:
 - specs pass

Change-Id: I8e5d903af8de01d12166b2f6ce861aec28720203
Reviewed-on: https://gerrit.instructure.com/15106
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Rob Orton <rob@instructure.com>